+++ Providers/EFCore/Revo.EFCore/DataAccess/Migrations/EFCoreMigrationsModule.cs
@@ -19,7 +19,7 @@ namespace Revo.EFCore.DataAccess.Migrations
         {
             Bind<IDatabaseMigrationProvider>()
                 .To<EFCoreDatabaseMigrationProvider>()
-                .InSingletonScope();
+                .InTaskScope();
             
             Bind<IMigrationScripterFactory>()
                 .To<MigrationScripterFactory>()