+++ Revo.Domain/ReadModel/EventEntityReadModel.cs
@@ -5,7 +5,7 @@
     /// version number for arbitrary concurrency control. This may prove useful if you need to update
     /// the read model based on other sources than just the event stream of its source aggregate.
     /// </summary>
-    public class EventEntityReadModel : EntityReadModel, IEventNumberVersioned
+    public abstract class EventEntityReadModel : EntityReadModel, IEventNumberVersioned
     {
         public int EventNumber { get; set; }
     }