+++ Revo.Infrastructure/DataAccess/Migrations/Events/DatabaseMigrationAppliedEvent.cs
@@ -1,0 +1,14 @@
+﻿using Revo.Core.Events;
+
+namespace Revo.Infrastructure.DataAccess.Migrations.Events
+{
+    public class DatabaseMigrationAppliedEvent : IEvent
+    {
+        public DatabaseMigrationAppliedEvent(DatabaseMigrationInfo migration)
+        {
+            Migration = migration;
+        }
+
+        public DatabaseMigrationInfo Migration { get; }
+    }
+}