+++ Revo.Infrastructure/DataAccess/Migrations/ExecuteDatabaseMigrationsCommand.cs
@@ -1,0 +1,8 @@
+﻿using Revo.Core.Commands;
+
+namespace Revo.Infrastructure.DataAccess.Migrations
+{
+    public class ExecuteDatabaseMigrationsCommand : ICommand
+    {
+    }
+}